[#23] 서비스 분리
