+++ src/main/kotlin/dev/weekend/slashcommand/presentation/CommandHandler.kt
@@ -1,6 +1,7 @@
 package dev.weekend.slashcommand.presentation
 
-import dev.weekend.slashcommand.application.CommandService
+import dev.weekend.slashcommand.application.BlindVoteService
+import dev.weekend.slashcommand.application.MbtiService
 import dev.weekend.slashcommand.presentation.model.MbtiInteractRequest
 import dev.weekend.slashcommand.presentation.model.MbtiTestRequest
 import dev.weekend.slashcommand.presentation.model.VoteCreateRequest
@@ -20,13 +21,14 @@ import org.springframework.web.reactive.function.server.bodyValueAndAwait
 
 @Component
 class CommandHandler(
-    private val commandService: CommandService,
+    private val blindVoteService: BlindVoteService,
+    private val mbtiService: MbtiService,
 ) {
     suspend fun createBlindVote(request: ServerRequest): ServerResponse {
         return withContext(Dispatchers.Default) {
             val createRequest = request.awaitBody<VoteCreateRequest>()
 
-            commandService.createBlindVote(createRequest)
+            blindVoteService.createBlindVote(createRequest)
                 .let { ok().bodyValueAndAwait(it) }
         }
     }
@@ -35,7 +37,7 @@ class CommandHandler(
         return withContext(Dispatchers.Default) {
             val interactRequest = request.awaitBody<VoteInteractRequest>()
 
-            commandService.interactBlindVote(interactRequest)
+            blindVoteService.interactBlindVote(interactRequest)
                 .let { ok().bodyValueAndAwait(it) }
         }
     }
@@ -44,7 +46,7 @@ class CommandHandler(
         return withContext(Dispatchers.Default) {
             val testRequest = request.awaitBody<MbtiTestRequest>()
 
-            commandService.testMbti(testRequest)
+            mbtiService.testMbti(testRequest)
                 .let { ok().bodyValueAndAwait(it) }
         }
     }
@@ -53,7 +55,7 @@ class CommandHandler(
         return withContext(Dispatchers.Default) {
             val interactRequest = request.awaitBody<MbtiInteractRequest>()
 
-            commandService.interactMbti(interactRequest)
+            mbtiService.interactMbti(interactRequest)
                 .let { ok().bodyValueAndAwait(it) }
         }
     }
